change cooldown (now a float from 0 to 1) and cooldownBars, remove unused train() method in abstractplayer

+++ source/world/drawables/train/wagons/weaponwagon.h
@@ -16,7 +16,7 @@ public:
 
     void setAimVector(const QVector3D& aimVector);
 
-    unsigned int maxCooldown() const override;
+    float maxCooldown() const override;
 
     WagonType wagonType() const override;
     float length() const override;
